requirements.txt
Flask==3.0.3
requests==2.32.3
boto3==1.34.162
python-dotenv==1.0.1
pytz==2024.1
gunicorn==23.0.0
cairosvg==2.7.1
